+++ CHANGELOG.md
@@ -17,7 +17,9 @@
 * add `SoftBreakAddsNewLinePlugin` plugin (`core` module)
 * `LinkResolverDef` defaults to `https` when a link does not have scheme information ([#75])
 * add `option` abstraction for `sample` module allowing switching of multiple cases in runtime via menu
+* non-empty bounds for `AsyncDrawable` when no dimensions are not yet available ([#189])
 
+[#189]: https://github.com/noties/Markwon/issues/189
 [#75]: https://github.com/noties/Markwon/issues/75
 [#204]: https://github.com/noties/Markwon/issues/204